+++ nodejs/doUtil.js
@@ -349,7 +349,7 @@ module.exports = {
      * is written to handle other partitions if they should enter the schema.
      *
      * @param {Object} declaration - The parsed declaration
-     * @param {Strint} classToFetch - The name of the class (DNS, VLAN, etc)
+     * @param {String} classToFetch - The name of the class (DNS, VLAN, etc)
      * @param {function} cb - Function to execute for each object. Will be called with 2 parameters
      *                        tenant and object declaration. Object declaration is the declaration
      *                        for just the object in question, not the whole declaration