+++ test/schema/analyticsSchemaTests.js
@@ -39,7 +39,7 @@ const validate = ajv.compile(analyticsSchema);
 describe('analytics.schema.json', () => {
     describe('Analytics', () => {
         describe('valid', () => {
-            it('should validate analytics data', () => {
+            it('should validate analytics data with IPv4 addresses for TCP protocol', () => {
                 const data = {
                     class: 'Analytics',
                     offboxEnabled: true,
@@ -49,18 +49,20 @@ describe('analytics.schema.json', () => {
                         '10.10.15.31'
                     ],
                     offboxTcpPort: 12345,
-                    debugEnabled: true
+                    debugEnabled: true,
+                    interval: 180
                 };
                 assert.ok(validate(data), getErrorString(validate));
             });
 
-            it('should also validate analytics data', () => {
+            it('should also validate analytics data with IPv6 addresses for TCP protocol', () => {
                 const data = {
                     class: 'Analytics',
                     offboxEnabled: true,
                     offboxProtocol: 'tcp',
                     offboxTcpAddresses: [
-                        'FE80:0000:0000:0000:0202:B3FF:FE1E:8330'
+                        'FE80:0000:0000:0000:0202:B3FF:FE1E:8330',
+                        'FE80:0000:0000:0000:0202:B3FF:FE1E:8331'
                     ],
                     offboxTcpPort: 12345,
                     debugEnabled: true,
@@ -69,19 +71,61 @@ describe('analytics.schema.json', () => {
                 assert.ok(validate(data), getErrorString(validate));
             });
 
-            it('should yet again validate analytics data', () => {
+            it('should yet again validate analytics data for HTTPS protocol', () => {
                 const data = {
                     class: 'Analytics',
                     offboxEnabled: true,
                     offboxProtocol: 'https',
+                    offboxTcpPort: 1245,
+                    offboxTcpAddresses: [
+                        'FE80:0000:0000:0000:0202:B3FF:FE1E:8330'
+                    ],
+                    sourceId: 'testSourceId',
                     debugEnabled: false
                 };
                 assert.ok(validate(data), getErrorString(validate));
             });
 
-            it('should validate analytics data one more time', () => {
+            it('should validate minimal analytics configuration', () => {
                 const data = {
                     class: 'Analytics'
+                };
+                assert.ok(validate(data), getErrorString(validate));
+            });
+
+            it('should validate analytics data without offbox* properties', () => {
+                const data = {
+                    class: 'Analytics',
+                    debugEnabled: true,
+                    interval: 180
+                };
+                assert.ok(validate(data), getErrorString(validate));
+            });
+
+            it('should validate analytics when all properties specified', () => {
+                const data = {
+                    class: 'Analytics',
+                    offboxEnabled: true,
+                    offboxProtocol: 'tcp',
+                    offboxTcpAddresses: [
+                        '10.10.15.30',
+                        '10.10.15.31',
+                        'FE80:0000:0000:0000:0202:B3FF:FE1E:8330'
+                    ],
+                    offboxTcpPort: 12345,
+                    sourceId: 'testSourceId',
+                    tenantId: 'testTenantId',
+                    debugEnabled: true,
+                    interval: 180
+                };
+                assert.ok(validate(data), getErrorString(validate));
+            });
+
+            it('should validate analytics data to disable configuration', () => {
+                const data = {
+                    class: 'Analytics',
+                    offboxTcpAddresses: [],
+                    offboxTcpPort: 0
                 };
                 assert.ok(validate(data), getErrorString(validate));
             });
@@ -91,32 +135,34 @@ describe('analytics.schema.json', () => {
             it('should invalidate illegal debugEnabled value', () => {
                 const data = {
                     class: 'Analytics',
-                    offboxEnabled: true,
                     debugEnabled: 'enabled'
                 };
                 assert.strictEqual(validate(data), false, 'illegal debugEnabled value should not be valied');
+                assert.strictEqual(/debugEnabled/.test(getErrorString(validate)), true);
             });
 
             it('should invalidate undersized interval value', () => {
                 const data = {
                     class: 'Analytics',
                     offboxEnabled: true,
-                    offboxProtocol: 'ecm-tm',
+                    offboxProtocol: 'tcp',
                     debugEnabled: true,
                     interval: 19
                 };
                 assert.strictEqual(validate(data), false, 'undersized interval value should not be valid');
+                assert.strictEqual(/interval/.test(getErrorString(validate)), true);
             });
 
             it('should invalidate oversized interval value', () => {
                 const data = {
                     class: 'Analytics',
                     offboxEnabled: true,
-                    offboxProtocol: 'ecm-tm',
+                    offboxProtocol: 'tcp',
                     debugEnabled: true,
                     interval: 301
                 };
                 assert.strictEqual(validate(data), false, 'oversized interval value should not be valid');
+                assert.strictEqual(/interval/.test(getErrorString(validate)), true);
             });
 
             it('should invalidate illegal offboxProtocol', () => {
@@ -126,6 +172,7 @@ describe('analytics.schema.json', () => {
                     offboxProtocol: 'udp'
                 };
                 assert.strictEqual(validate(data), false, 'illegal offboxProtocol value should not be valid');
+                assert.strictEqual(/offboxProtocol/.test(getErrorString(validate)), true);
             });
 
             it('should invalidate illegal offboxTcpAddresses value', () => {
@@ -153,6 +200,7 @@ describe('analytics.schema.json', () => {
                     offboxTcpPort: 65536
                 };
                 assert.strictEqual(validate(data), false, 'out of range offbox port should not be valid');
+                assert.strictEqual(/offboxTcpPort/.test(getErrorString(validate)), true);
             });
 
             it('should invalidate illegal offboxEnabled value', () => {
@@ -161,6 +209,7 @@ describe('analytics.schema.json', () => {
                     offboxEnabled: 'disabled'
                 };
                 assert.strictEqual(validate(data), false, 'illegal offboxEnabled value should not be valid');
+                assert.strictEqual(/offboxEnabled/.test(getErrorString(validate)), true);
             });
 
             it('should invalidate missing offboxProtocol value when offboxEnabled true', () => {
@@ -169,6 +218,7 @@ describe('analytics.schema.json', () => {
                     offboxEnabled: true
                 };
                 assert.strictEqual(validate(data), false, 'missing offboxProtocol value should not be valid');
+                assert.strictEqual(/offboxProtocol/.test(getErrorString(validate)), true);
             });
 
             it('should invalidate missing offboxTcpPort when offboxProtocol is tcp', () => {
@@ -182,6 +232,7 @@ describe('analytics.schema.json', () => {
                 };
                 assert.strictEqual(validate(data), false,
                     'missing offboxTcpPort with tcp protocol should not be valid');
+                assert.strictEqual(/offboxTcpPort/.test(getErrorString(validate)), true);
             });
 
             it('should invalidate missing offboxTcpAddresses when offboxProtocol tcp', () => {
@@ -193,6 +244,7 @@ describe('analytics.schema.json', () => {
                 };
                 assert.strictEqual(validate(data), false,
                     'missing offboxTcpAddresses with tcp protocol should not be valid');
+                assert.strictEqual(/offboxTcpAddresses/.test(getErrorString(validate)), true);
             });
 
             it('should invalidate additional properties', () => {
@@ -201,6 +253,75 @@ describe('analytics.schema.json', () => {
                     enabled: true
                 };
                 assert.strictEqual(validate(data), false, 'additional properties hsould not be valid');
+                assert.strictEqual(/additional/.test(getErrorString(validate)), true);
+            });
+
+            it('should invalidate missing offboxTcpAddresses, offboxTcpPort when offboxProtocol tcp', () => {
+                const data = {
+                    class: 'Analytics',
+                    offboxEnabled: true,
+                    offboxProtocol: 'tcp'
+                };
+                assert.strictEqual(validate(data), false,
+                    'missing offboxTcpAddresses, offboxTcpPort with tcp protocol should not be valid');
+
+                const errorMsg = getErrorString(validate);
+                assert.strictEqual(/offboxTcpAddresses/.test(errorMsg) || /offboxTcpPort/.test(errorMsg), true);
+            });
+
+            it('should invalidate missing offboxTcpAddresses, offboxProtocol when offboxTcpPort is not 0', () => {
+                const data = {
+                    class: 'Analytics',
+                    offboxEnabled: true,
+                    offboxTcpPort: 1
+                };
+                assert.strictEqual(validate(data), false,
+                    'missing offboxTcpAddresses, offboxProtocol when offboxTcpPort is not 0 should not be valid');
+
+                const errorMsg = getErrorString(validate);
+                assert.strictEqual(/offboxTcpAddresses/.test(errorMsg) || /offboxProtocol/.test(errorMsg), true);
+            });
+
+            it('should invalidate missing offboxTcpPort, offboxProtocol when offboxTcpAddresses is not empty', () => {
+                const data = {
+                    class: 'Analytics',
+                    offboxEnabled: true,
+                    offboxTcpAddresses: [
+                        '10.0.0.10'
+                    ]
+                };
+                assert.strictEqual(validate(data), false,
+                    'missing offboxProtocol, offboxTcpPort when offboxTcpAddresses is not empty should not be valid');
+
+                const errorMsg = getErrorString(validate);
+                assert.strictEqual(/offboxTcpPort/.test(errorMsg) || /offboxProtocol/.test(errorMsg), true);
+            });
+
+            it('should invalidate missing offboxTcpAddresses, offboxTcpPort when offboxProtocol https', () => {
+                const data = {
+                    class: 'Analytics',
+                    offboxEnabled: true,
+                    offboxProtocol: 'https'
+                };
+                assert.strictEqual(validate(data), false,
+                    'missing offboxTcpAddresses, offboxTcpPort with https protocol should not be valid');
+
+                const errorMsg = getErrorString(validate);
+                assert.strictEqual(/offboxTcpAddresses/.test(errorMsg) || /offboxTcpPort/.test(errorMsg), true);
+            });
+
+            it('should invalidate missing sourceId when offboxProtocol https', () => {
+                const data = {
+                    class: 'Analytics',
+                    offboxEnabled: true,
+                    offboxProtocol: 'https',
+                    offboxTcpPort: 12345,
+                    offboxTcpAddresses: [
+                        '10.0.0.10'
+                    ]
+                };
+                assert.strictEqual(validate(data), false, 'missing sourceId with https protocol should not be valid');
+                assert.strictEqual(/sourceId/.test(getErrorString(validate)), true);
             });
         });
     });